--- a/.gitignore
+++ b/.gitignore
@@ -6,3 +6,4 @@ docs/
 .nyc_output/coverage-final.json
 .vscode/settings.json
 benchmark_data/
+.vscode/