--- a/src/message/rpc.js
+++ b/src/message/rpc.js
@@ -1,3 +1,0 @@
-import cjs from "./rpc.cjs"
-
-export const {RPC} = cjs