--- a/src/types.ts
+++ b/src/types.ts
@@ -16,7 +16,7 @@ export interface AddrInfo {
  * Compute a local non-spec'ed msg-id for faster de-duplication of seen messages.
  * Used exclusively for a local seen_cache
  */
-export interface FastMsgIdFn { (msg: RPC.IMessage): string | number }
+export interface FastMsgIdFn { (msg: RPC.Message): string | number }
 
 /**
  * By default, gossipsub only provide a browser friendly function to convert Uint8Array message id to string.