**Motivation**
Switch from protobufjs to protons

**Description**

While decode flow is comparable or a little bit faster, encode flow is ~1.5x slower

```
  protobuf
    ✔ decode Attestation message using protobufjs                          3318731 ops/s    301.3200 ns/op       396763 runs    120 s
    ✔ encode Attestation message using protobufjs                          2515603 ops/s    397.5190 ns/op       300351 runs    120 s
```

```
  protobuf
    ✔ decode Attestation message using protons-runtime 5.1.0               3996387 ops/s    250.2260 ns/op        -     478012 runs    120 s
    ✔ encode Attestation message using protons-runtime 5.1.0               1667211 ops/s    599.8040 ns/op        -     199077 runs    120 s
```

cc @wemeetagain 

Closes #453 
Closes #318

**TODOs**
Test with lodestar to see how it effect the performance