@tuyennhv 2024-01-31
sorry @achingbrain for the delay, I just rebased against master, testing it with lodestar now cc @wemeetagain