@tuyennhv 2024-02-01
somehow rss in this branch is consistently higher than with gossipsub v11.1.0 when I test on different nodes

| # | this branch | v11.1.0 |
|--|-------------|-------|
|holesky 1k|12.5GB|9GB|
|mainnet|9.5GB|7GB|
|holesky 64|15GB|10GB|