@achingbrain 2024-02-02
That's interesting, protons is synchronous and stateless so any memory it uses would be garbage collectable immediately.

That said, it's going to have a different memory usage pattern to `protobuf.js` so it could in turn affect how other components use memory.

Do you have a view on what is taking up the extra space?