@tuyennhv 2024-02-06
> That's interesting, protons is synchronous and stateless so any memory it uses would be garbage collectable immediately.

@achingbrain  how is it different from `protobuf.js`? I thought `protobuf.js` is synchronous and stateless too

> That said, it's going to have a different memory usage pattern to `protobuf.js` so it could in turn affect how other components use memory.
>
> Do you have a view on what is taking up the extra space?

I think it's worth to compare https://github.com/protobufjs/protobuf.js/blob/master/src/reader.js and https://github.com/ipfs/protons/blob/main/packages/protons-runtime/src/utils/reader.ts . From a quick look it shares a lot of similarities