@achingbrain 2024-02-06
> how is it different from protobuf.js

It isn't, the reader/writer parts of protons are from protobuf.js, ported to ESM.

> From a quick look it shares a lot of similarities

Yeah, it's basically the same code which is why it's surprising the RSS usage is higher.

For the same workload the memory behaviour should be comparable.  I don't have a view on the testing you've been doing with Lodestar so I can only assume the workload is different.

Would it be possible for you to record the inputs and outputs of the protobuf.js usage in the app, (e.g. write the buffers being decoded/objects being encoded out to a file), then we could replay them without the rest of lodestar present and observe the memory usage?