@tuyennhv 2024-02-08
mesh peers are similar to `v11.x.x` in lodestar, ready for review

<img width="1331" alt="Screenshot 2024-02-09 at 06 36 20" src="https://github.com/ChainSafe/js-libp2p-gossipsub/assets/10568965/cac33e31-e261-45cf-9cbf-cf9701e993c3">