@tuyennhv 2024-02-10
I've tested this branch for ~18h: mesh peers are good (since ~6h30)

<img width="834" alt="Screenshot 2024-02-10 at 07 21 13" src="https://github.com/ChainSafe/js-libp2p-gossipsub/assets/10568965/5bdced59-0188-4aba-bd6f-44fc58b264f4">

external memory is unstable, not sure if it's an issue, just something to notice for later reference

<img width="826" alt="Screenshot 2024-02-10 at 07 22 15" src="https://github.com/ChainSafe/js-libp2p-gossipsub/assets/10568965/75ad92b8-9025-4847-977d-58942e213371">